+++ src/sdk/PnP.Core/Model/SharePoint/Core/Internal/File.cs
@@ -413,46 +413,59 @@ namespace PnP.Core.Model.SharePoint
 
             if (!string.IsNullOrEmpty(response.Json))
             {
-                var document = JsonSerializer.Deserialize<JsonElement>(response.Json);
-                if (document.TryGetProperty("d", out JsonElement root))
+                return ProcessRecycleResponse(response.Json);
+            }
+
+            return Guid.Empty;
+        }
+
+        private static Guid ProcessRecycleResponse(string json)
+        {
+            var document = JsonSerializer.Deserialize<JsonElement>(json);
+            if (document.TryGetProperty("d", out JsonElement root))
+            {
+                if (root.TryGetProperty("Recycle", out JsonElement recycleBinItemId))
                 {
-                    if (root.TryGetProperty("Recycle", out JsonElement recycleBinItemId))
-                    {
-                        // return the recyclebin item id
-                        return recycleBinItemId.GetGuid();
-                    }
+                    // return the recyclebin item id
+                    return recycleBinItemId.GetGuid();
                 }
             }
 
             return Guid.Empty;
         }
 
-        public void RecycleBatch()
+        public IBatchSingleResult<BatchResultValue<Guid>> RecycleBatch()
         {
-            RecycleBatchAsync().GetAwaiter().GetResult();
+            return RecycleBatchAsync().GetAwaiter().GetResult();
         }
 
-        public void RecycleBatch(Batch batch)
+        public IBatchSingleResult<BatchResultValue<Guid>> RecycleBatch(Batch batch)
         {
-            RecycleBatchAsync(batch).GetAwaiter().GetResult();
+            return RecycleBatchAsync(batch).GetAwaiter().GetResult();
         }
 
-        public async Task RecycleBatchAsync()
+        public async Task<IBatchSingleResult<BatchResultValue<Guid>>> RecycleBatchAsync()
         {
-            await RecycleBatchAsync(PnPContext.CurrentBatch).ConfigureAwait(false);
+            return await RecycleBatchAsync(PnPContext.CurrentBatch).ConfigureAwait(false);
         }
 
-        public async Task RecycleBatchAsync(Batch batch)
+        public async Task<IBatchSingleResult<BatchResultValue<Guid>>> RecycleBatchAsync(Batch batch)
         {
             var entity = EntityManager.GetClassInfo(GetType(), this);
             string recycleEndpointUrl = $"{entity.SharePointUri}/recycle";
 
             var apiCall = new ApiCall(recycleEndpointUrl, ApiType.SPORest)
             {
-                RemoveFromModel = true
+                RemoveFromModel = true,
+                RawSingleResult = new BatchResultValue<Guid>(Guid.Empty),
+                RawResultsHandler = (json, apiCall) =>
+                {
+                    (apiCall.RawSingleResult as BatchResultValue<Guid>).Value = ProcessRecycleResponse(json);
+                }
             };
 
-            await RawRequestBatchAsync(batch, apiCall, HttpMethod.Post).ConfigureAwait(false);
+            var batchRequest = await RawRequestBatchAsync(batch, apiCall, HttpMethod.Post).ConfigureAwait(false);
+            return new BatchSingleResult<BatchResultValue<Guid>>(batch, batchRequest.Id, apiCall.RawSingleResult as BatchResultValue<Guid>);
         }
         #endregion
 
@@ -650,7 +663,56 @@ namespace PnP.Core.Model.SharePoint
             var apiResult = await RawRequestAsync(apiCall, HttpMethod.Post).ConfigureAwait(false);
 
             // Process response
-            var root = JsonDocument.Parse(apiResult.Json).RootElement.GetProperty("d");
+            return ProcessClassifyAndExtractResponse(apiResult.Json);
+        }
+
+        public ISyntexClassifyAndExtractResult ClassifyAndExtractFile()
+        {
+            return ClassifyAndExtractFileAsync().GetAwaiter().GetResult();
+        }
+
+        public async Task<IBatchSingleResult<ISyntexClassifyAndExtractResult>> ClassifyAndExtractFileBatchAsync(Batch batch)
+        {
+            ApiCall apiCall = CreateClassifyAndExtractApiCall();
+            apiCall.RawSingleResult = new SyntexClassifyAndExtractResult();
+            apiCall.RawResultsHandler = (json, apiCall) =>
+            {
+                var result = ProcessClassifyAndExtractResponse(json);
+                (apiCall.RawSingleResult as SyntexClassifyAndExtractResult).Created = result.Created;
+                (apiCall.RawSingleResult as SyntexClassifyAndExtractResult).DeliverDate = result.DeliverDate;
+                (apiCall.RawSingleResult as SyntexClassifyAndExtractResult).ErrorMessage = result.ErrorMessage;
+                (apiCall.RawSingleResult as SyntexClassifyAndExtractResult).Id = result.Id;
+                (apiCall.RawSingleResult as SyntexClassifyAndExtractResult).Status = result.Status;
+                (apiCall.RawSingleResult as SyntexClassifyAndExtractResult).StatusCode = result.StatusCode;
+                (apiCall.RawSingleResult as SyntexClassifyAndExtractResult).TargetServerRelativeUrl = result.TargetServerRelativeUrl;
+                (apiCall.RawSingleResult as SyntexClassifyAndExtractResult).TargetSiteUrl = result.TargetSiteUrl;
+                (apiCall.RawSingleResult as SyntexClassifyAndExtractResult).TargetWebServerRelativeUrl = result.TargetWebServerRelativeUrl;
+                (apiCall.RawSingleResult as SyntexClassifyAndExtractResult).WorkItemType = result.WorkItemType;
+            };
+
+            var batchRequest = await RawRequestBatchAsync(batch, apiCall, HttpMethod.Post).ConfigureAwait(false);
+            
+            return new BatchSingleResult<ISyntexClassifyAndExtractResult>(batch, batchRequest.Id, apiCall.RawSingleResult as ISyntexClassifyAndExtractResult);
+        }
+
+        public IBatchSingleResult<ISyntexClassifyAndExtractResult> ClassifyAndExtractFileBatch(Batch batch)
+        {
+            return ClassifyAndExtractFileBatchAsync(batch).GetAwaiter().GetResult();
+        }
+
+        public async Task<IBatchSingleResult<ISyntexClassifyAndExtractResult>> ClassifyAndExtractFileBatchAsync()
+        {
+            return await ClassifyAndExtractFileBatchAsync(PnPContext.CurrentBatch).ConfigureAwait(false);
+        }
+
+        public IBatchSingleResult<ISyntexClassifyAndExtractResult> ClassifyAndExtractFileBatch()
+        {
+            return ClassifyAndExtractFileBatchAsync().GetAwaiter().GetResult();
+        }
+
+        private static ISyntexClassifyAndExtractResult ProcessClassifyAndExtractResponse(string json)
+        {
+            var root = JsonDocument.Parse(json).RootElement.GetProperty("d");
             return new SyntexClassifyAndExtractResult
             {
                 Created = root.GetProperty("Created").GetDateTime(),
@@ -664,32 +726,6 @@ namespace PnP.Core.Model.SharePoint
                 TargetSiteUrl = root.GetProperty("TargetSiteUrl").GetString(),
                 TargetWebServerRelativeUrl = root.GetProperty("TargetWebServerRelativeUrl").GetString()
             };
-        }
-
-        public ISyntexClassifyAndExtractResult ClassifyAndExtractFile()
-        {
-            return ClassifyAndExtractFileAsync().GetAwaiter().GetResult();
-        }
-
-        public async Task ClassifyAndExtractFileBatchAsync(Batch batch)
-        {
-            ApiCall apiCall = CreateClassifyAndExtractApiCall();
-            await RawRequestBatchAsync(batch, apiCall, HttpMethod.Post).ConfigureAwait(false);
-        }
-
-        public void ClassifyAndExtractFileBatch(Batch batch)
-        {
-            ClassifyAndExtractFileBatchAsync(batch).GetAwaiter().GetResult();
-        }
-
-        public async Task ClassifyAndExtractFileBatchAsync()
-        {
-            await ClassifyAndExtractFileBatchAsync(PnPContext.CurrentBatch).ConfigureAwait(false);
-        }
-
-        public void ClassifyAndExtractFileBatch()
-        {
-            ClassifyAndExtractFileBatchAsync().GetAwaiter().GetResult();
         }
 
         private ApiCall CreateClassifyAndExtractApiCall()