+++ src/sdk/PnP.Core/Services/Core/ApiCall.cs
@@ -1,4 +1,7 @@
-﻿namespace PnP.Core.Services
+﻿using System;
+using System.Collections;
+
+namespace PnP.Core.Services
 {
     /// <summary>
     /// Struct that defines the API call to make
@@ -13,6 +16,9 @@
             XmlBody = null;
             ReceivingProperty = receivingProperty;
             RawRequest = false;
+            RawSingleResult = null;
+            RawEnumerableResult = null;
+            RawResultsHandler = null;
             Commit = false;
             Interactive = false;
             BinaryBody = null;
@@ -31,6 +37,9 @@
             XmlBody = xmlBody;
             ReceivingProperty = receivingProperty;
             RawRequest = false;
+            RawSingleResult = null;
+            RawEnumerableResult = null;
+            RawResultsHandler = null;
             Commit = false;
             Interactive = false;
             BinaryBody = null;
@@ -74,6 +83,21 @@
         /// Is this a raw request that does not require automatic parsing of the returned json?
         /// </summary>
         internal bool RawRequest { get; set; }
+
+        /// <summary>
+        /// List containing the parsed Raw results (if a collection is returned)
+        /// </summary>
+        internal IEnumerable RawEnumerableResult { get; set; }
+
+        /// <summary>
+        /// Result object from this API call (if a single result is returned)
+        /// </summary>
+        internal object RawSingleResult { get; set; }
+
+        /// <summary>
+        /// Event handler triggered to parse raw results
+        /// </summary>
+        internal Action<string, ApiCall> RawResultsHandler { get; set; }
 
         /// <summary>
         /// When set to true the current model item will be committed as changes are synced with the server